+++ spec/requests/edit_collection_spec.rb
@@ -51,7 +51,7 @@ RSpec.describe 'Edit collection' do
     before do
       allow(Sdr::Repository).to receive(:find).with(druid: collection.druid).and_return(cocina_object)
       allow(Sdr::Repository).to receive(:status).with(druid: collection.druid).and_return(version_status)
-      allow(ToCollectionForm::RoundtripValidator).to receive(:call)
+      allow(CollectionRoundtripper).to receive(:call)
 
       sign_in(user)
     end
@@ -63,7 +63,7 @@ RSpec.describe 'Edit collection' do
 
       follow_redirect!
       expect(response.body).to include('This collection cannot be edited.')
-      expect(ToCollectionForm::RoundtripValidator).not_to have_received(:call)
+      expect(CollectionRoundtripper).not_to have_received(:call)
     end
   end
 
@@ -78,7 +78,7 @@ RSpec.describe 'Edit collection' do
     before do
       allow(Sdr::Repository).to receive(:find).with(druid: collection.druid).and_return(cocina_object)
       allow(Sdr::Repository).to receive(:status).with(druid: collection.druid).and_return(version_status)
-      allow(ToCollectionForm::RoundtripValidator).to receive(:call).and_call_original
+      allow(CollectionRoundtripper).to receive(:call).and_call_original
 
       sign_in(user)
     end
@@ -90,7 +90,7 @@ RSpec.describe 'Edit collection' do
 
       follow_redirect!
       expect(response.body).to include('This collection cannot be edited.')
-      expect(ToCollectionForm::RoundtripValidator).to have_received(:call)
+      expect(CollectionRoundtripper).to have_received(:call)
     end
   end
 end
